requirements.txt
Django==1.8.2
django-adminfiles==1.0.1
django-markitup==2.2.2
djangoembed==0.1.1
gdata==2.0.18
Markdown==2.6.2
oauthlib==0.7.2
Pillow==2.8.2
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
six==1.9.0
sorl-thumbnail==12.2
wheel==0.24.0
